Renamed provider to virtualbox
Looking at other providers, this seems to be the the usual approach.
