+++ virtualbox/image.go
@@ -1,4 +1,4 @@
-package provider
+package virtualbox
 
 import (
 	"crypto/md5"
