+++ virtualbox/image_test.go
@@ -1,4 +1,4 @@
-package provider
+package virtualbox
 
 import (
 	"io/ioutil"
